Finished calc/scenario_test.py
