+++ openquake/hazardlib/__init__.py
@@ -20,147 +20,7 @@ hazardlib stands for Hazard Library.
 """
 import operator
 import itertools
+import unittest
 import collections
 import numpy
-from openquake.baselib import __version__, InvalidFile, hdf5, general
-from openquake.hazardlib import (
-    calc, geo, gsim, mfd, scalerel, source, const, correlation, imt, pmf, site,
-    tom, nrml, near_fault, sourceconverter, gsim_lt, contexts)
-from openquake.hazardlib.source.rupture import (
-    EBRupture, get_ruptures, _get_rupture)
-from openquake.hazardlib.calc.filters import MagDepDistance
-
-bytrt = operator.attrgetter('tectonic_region_type')
-Input = collections.namedtuple('Input', 'model sitecol gsim_lt cmakerdict')
-
-
-def _get_site_model(fname, req_site_params):
-    sm = hdf5.read_csv(fname, site.site_param_dt).array
-    sm['lon'] = numpy.round(sm['lon'], 5)
-    sm['lat'] = numpy.round(sm['lat'], 5)
-    dupl = general.get_duplicates(sm, 'lon', 'lat')
-    if dupl:
-        raise InvalidFile(
-            'Found duplicate sites %s in %s' % (dupl, fname))
-    z = numpy.zeros(len(sm), sorted(sm.dtype.descr))
-    for name in z.dtype.names:
-        z[name] = sm[name]
-    return z
-
-
-def _get_sitecol(hparams, req_site_params):
-    """
-    :param hparams: a dictionary of hazard parameters
-    """
-    if 'sites' in hparams:
-        sm = None
-        mesh = geo.Mesh.from_coords(hparams['sites'])
-    elif 'site_model_file' in hparams:
-        sm = _get_site_model(hparams['site_mode_file'], req_site_params)
-        mesh = geo.Mesh(sm['lon'], sm['lat'])
-    else:
-        raise KeyError('Missing sites or site_mode_file')
-    return site.SiteCollection.from_points(
-        mesh.lons, mesh.lats, mesh.depths, sm or hparams, req_site_params)
-
-
-def _rupture_groups(ebruptures):
-    ebruptures.sort(key=bytrt)
-    rup_groups = []
-    for trt, ebrs in itertools.groupby(ebruptures, bytrt):
-        rup_groups.append(sourceconverter.SourceGroup(trt, ebrs))
-    return rup_groups
-
-
-def _get_ebruptures(fname, conv=None, ses_seed=None):
-    """
-    :param fname: path to a rupture file (XML or CSV)
-    :param conv: RuptureConverter instanc, used for XML ruptures
-    :param ses_seed: used for XML ruptures
-    :returns: a list of one or more EBRuptures
-    """
-    if fname.endswith('.xml'):
-        [rup_node] = nrml.read(fname)
-        rup = conv.convert_node(rup_node)
-        rup.tectonic_region_type = '*'  # no TRT for scenario ruptures
-        rup.rup_id = ses_seed
-        ebrs = [EBRupture(rup, 'NA', 0, id=rup.rup_id, scenario=True)]
-        return ebrs
-
-    assert fname.endswith('.csv'), fname
-    aw = get_ruptures(fname)
-    ebrs = []
-    for i, rec in enumerate(aw.array):
-        rupture = _get_rupture(rec, aw.geoms[i], aw.trts[rec['trt_smr']])
-        ebr = EBRupture(rupture, rec['source_id'], rec['trt_smr'],
-                        rec['n_occ'], rec['id'], rec['e0'])
-        ebrs.append(ebr)
-    return ebrs
-
-
-def read_input(hparams):
-    """
-    :param hparams: a dictionary of hazard parameters
-    :returns: an Input tuple (groups, sitecol, gsim_lt, cmakers)
-
-    The dictionary must contain the keys
-
-    - "maximum_distance"
-    - "imts" or "imtls"
-    - "source_model_file" or "rupture_model_file"
-
-    Moreover:
-
-    - if source_model_file "investigation_time" is mandatory
-    - if rupture_model_file "number_of_ground_motion_fields" and "ses_seed"
-      are mandatory
-
-    The optional keys include
-
-    - "rupture_mesh_spacing" (default 5.)
-    - "complex_fault_mesh_spacing" (default rupture_mesh_spacing)
-    - "width_of_mfd_bin" (default 1.)
-    - "area_source_discretization"
-    - "minimum_magnitude"
-    - "discard_trts" (default "")
-    """
-    assert 'imts' in hparams or 'imtls' in hparams
-    assert isinstance(hparams['maximum_distance'], MagDepDistance)
-    smfname = hparams.get('source_model_file')
-    if smfname:  # nonscenario
-        assert 'investigation_time' in hparams
-    rmfname = hparams.get('rupture_model_file')
-    if rmfname:
-        ngmfs = hparams["number_of_ground_motion_fields"]
-        ses_seed = hparams["ses_seed"]
-    converter = sourceconverter.SourceConverter(
-        hparams.get('investigation_time', 50.),
-        hparams.get('rupture_mesh_spacing', 5.),
-        hparams.get('complex_fault_mesh_spacing'),
-        hparams.get('width_of_mfd_bin', 1.0),
-        hparams.get('area_source_discretization'),
-        hparams.get('minimum_magnitude', {'default': 0}),
-        hparams.get('source_id'),
-        discard_trts=hparams.get('discard_trts', ''))
-    if smfname:
-        [sm] = nrml.read_source_models([smfname], converter)
-        groups = sm.src_groups
-    elif rmfname:
-        ebrs = _get_ebruptures(rmfname, converter, ses_seed)
-        groups = _rupture_groups(ebrs)
-    else:
-        raise KeyError('Missing source_model_file or rupture_file')
-    trts = set(grp.trt for grp in groups)
-    if 'gsim' in hparams:
-        lt = gsim_lt.GsimLogicTree.from_(hparams['gsim'])
-    else:
-        lt = gsim_lt.GsimLogicTree(hparams['gsim_logic_tree_file'], trts)
-    cmaker = {}  # trt => cmaker
-    for trt, rlzs_by_gsim in lt.get_rlzs_by_gsim_trt().items():
-        cmaker[trt] = contexts.ContextMaker(trt, rlzs_by_gsim, hparams)
-    if rmfname and rmfname.endswith('.xml'):
-        for grp in groups:
-            for ebr in grp:
-                ebr.n_occ = ngmfs * len(rlzs_by_gsim)
-    sitecol = _get_sitecol(hparams, lt.req_site_params)
-    return Input(groups, sitecol, lt, cmaker)
+from openquake.baselib import __version__, InvalidFile
