+++ openquake/hazardlib/calc/__init__.py
@@ -20,7 +20,160 @@
 Package :mod:`openquake.hazardlib.calc` contains hazard calculator modules
 and utilities for them, such as :mod:`~openquake.hazardlib.calc.filters`.
 """
-from openquake.hazardlib.calc.gmf import ground_motion_fields
-# from disagg we want to import main calc function
-# as well as all the pmf extractors
-from openquake.hazardlib.calc.disagg import *
+import operator
+import itertools
+import unittest
+import collections
+import numpy
+from openquake.baselib import hdf5, general, InvalidFile
+from openquake.hazardlib import (
+    geo, site, nrml, sourceconverter, gsim_lt, contexts)
+from openquake.hazardlib.source.rupture import (
+    EBRupture, get_ruptures, _get_rupture)
+from openquake.hazardlib.calc.filters import MagDepDistance
+
+bytrt = operator.attrgetter('tectonic_region_type')
+Input = collections.namedtuple('Input', 'groups sitecol gsim_lt cmakerdict')
+
+
+def _get_site_model(fname, req_site_params):
+    sm = hdf5.read_csv(fname, site.site_param_dt).array
+    sm['lon'] = numpy.round(sm['lon'], 5)
+    sm['lat'] = numpy.round(sm['lat'], 5)
+    dupl = general.get_duplicates(sm, 'lon', 'lat')
+    if dupl:
+        raise InvalidFile(
+            'Found duplicate sites %s in %s' % (dupl, fname))
+    z = numpy.zeros(len(sm), sorted(sm.dtype.descr))
+    for name in z.dtype.names:
+        z[name] = sm[name]
+    return z
+
+
+def _get_sitecol(hparams, req_site_params):
+    """
+    :param hparams: a dictionary of hazard parameters
+    """
+    if 'sites' in hparams:
+        sm = unittest.mock.Mock(**hparams)
+        mesh = geo.Mesh.from_coords(hparams['sites'])
+    elif 'site_model_file' in hparams:
+        sm = _get_site_model(hparams['site_mode_file'], req_site_params)
+        mesh = geo.Mesh(sm['lon'], sm['lat'])
+    else:
+        raise KeyError('Missing sites or site_mode_file')
+    return site.SiteCollection.from_points(
+        mesh.lons, mesh.lats, mesh.depths, sm, req_site_params)
+
+
+def _rupture_groups(ebruptures):
+    ebruptures.sort(key=bytrt)
+    rup_groups = []
+    for trt, ebrs in itertools.groupby(ebruptures, bytrt):
+        grp = sourceconverter.SourceGroup(trt)
+        grp.sources = list(ebrs)
+        rup_groups.append(grp)
+    return rup_groups
+
+
+def _get_ebruptures(fname, conv=None, ses_seed=None):
+    """
+    :param fname: path to a rupture file (XML or CSV)
+    :param conv: RuptureConverter instanc, used for XML ruptures
+    :param ses_seed: used for XML ruptures
+    :returns: a list of one or more EBRuptures
+    """
+    if fname.endswith('.xml'):
+        [rup_node] = nrml.read(fname)
+        rup = conv.convert_node(rup_node)
+        rup.tectonic_region_type = '*'  # no TRT for scenario ruptures
+        rup.rup_id = ses_seed
+        ebrs = [EBRupture(rup, 'NA', 0, id=rup.rup_id, scenario=True)]
+        return ebrs
+
+    assert fname.endswith('.csv'), fname
+    aw = get_ruptures(fname)
+    ebrs = []
+    for i, rec in enumerate(aw.array):
+        rupture = _get_rupture(rec, aw.geoms[i], aw.trts[rec['trt_smr']])
+        ebr = EBRupture(rupture, rec['source_id'], rec['trt_smr'],
+                        rec['n_occ'], rec['id'], rec['e0'])
+        ebrs.append(ebr)
+    return ebrs
+
+
+def read_input(hparams):
+    """
+    :param hparams: a dictionary of hazard parameters
+    :returns: an Input tuple (groups, sitecol, gsim_lt, cmakers)
+
+    The dictionary must contain the keys
+
+    - "maximum_distance"
+    - "imtls"
+    - "source_model_file" or "rupture_model_file"
+    - "sites" or "site_model_file"
+
+    Moreover:
+
+    - if "source_model_file" is given, then "investigation_time" is mandatory
+    - if "rupture_model_file" is given, the "number_of_ground_motion_fields"
+      and "ses_seed" are mandatory
+    - if there is an area source, then "area_source_discretization" is needed
+    - if  "site_model_file" is missing, then global site parameters are needed
+
+    The optional keys include
+
+    - "rupture_mesh_spacing" (default 5.)
+    - "complex_fault_mesh_spacing" (default rupture_mesh_spacing)
+    - "width_of_mfd_bin" (default 1.)
+    - "minimum_magnitude"
+    - "discard_trts" (default "")
+    - "number_of_logic_tree_samples" (default 0)
+    """
+    assert 'imts' in hparams or 'imtls' in hparams
+    assert isinstance(hparams['maximum_distance'], MagDepDistance)
+    smfname = hparams.get('source_model_file')
+    if smfname:  # nonscenario
+        itime = hparams['investigation_time']
+    else:
+        itime = 50.  # ignored in scenario
+    rmfname = hparams.get('rupture_model_file')
+    if rmfname:
+        ngmfs = hparams["number_of_ground_motion_fields"]
+        ses_seed = hparams["ses_seed"]
+    converter = sourceconverter.SourceConverter(
+        itime,
+        hparams.get('rupture_mesh_spacing', 5.),
+        hparams.get('complex_fault_mesh_spacing'),
+        hparams.get('width_of_mfd_bin', 1.0),
+        hparams.get('area_source_discretization'),
+        hparams.get('minimum_magnitude', {'default': 0}),
+        hparams.get('source_id'),
+        discard_trts=hparams.get('discard_trts', ''))
+    if smfname:
+        [sm] = nrml.read_source_models([smfname], converter)
+        groups = sm.src_groups
+    elif rmfname:
+        ebrs = _get_ebruptures(rmfname, converter, ses_seed)
+        groups = _rupture_groups(ebrs)
+    else:
+        raise KeyError('Missing source_model_file or rupture_file')
+    trts = set(grp.trt for grp in groups)
+    if 'gsim' in hparams:
+        lt = gsim_lt.GsimLogicTree.from_(hparams['gsim'])
+    else:
+        lt = gsim_lt.GsimLogicTree(hparams['gsim_logic_tree_file'], trts)
+    cmaker = {}  # trt => cmaker
+    for trt, rlzs_by_gsim in lt.get_rlzs_by_gsim_trt().items():
+        cmaker[trt] = contexts.ContextMaker(trt, rlzs_by_gsim, hparams)
+    if rmfname:
+        # for instance for 2 TRTs with 5x1 GSIMs and ngmfs=10, then the
+        # numbers of occupation are 50 and 10 respectively, for a total
+        # of 60 events, see scenario/case_13
+        for grp in groups:
+            ngsims = len(cmaker[grp.trt].gsims)
+            for ebr in grp:
+                ebr.n_occ = ngmfs * ngsims
+    sitecol = _get_sitecol(hparams, lt.req_site_params)
+    return Input(groups, sitecol, lt, cmaker)
